APP/UI/requirements.txt
streamlit==1.31.0
# psycopg2==2.9.9
psycopg2-binary
